+++ mohaeng/src/main/java/com/mohaeng/notification/domain/model/strategy/DeleteApplicationFormBecauseClubIsDeletedNotificationMakeStrategy.java
@@ -1,10 +1,10 @@
 package com.mohaeng.notification.domain.model.strategy;
 
-import com.mohaeng.applicationform.domain.event.DeleteApplicationFormEvent;
 import com.mohaeng.notification.domain.model.Notification;
 import com.mohaeng.notification.domain.model.NotificationMakeStrategy;
 import com.mohaeng.notification.domain.model.kind.DeleteApplicationFormBecauseClubIsDeletedNotification;
 import com.mohaeng.notification.domain.model.value.Receiver;
+import com.mohaeng.participant.domain.event.DeleteApplicationFormEvent;
 import org.springframework.stereotype.Component;
 
 import java.util.List;
