+++ mohaeng/src/main/java/com/mohaeng/notification/domain/model/strategy/OfficerRejectClubJoinApplicationNotificationMakeStrategy.java
@@ -1,10 +1,10 @@
 package com.mohaeng.notification.domain.model.strategy;
 
-import com.mohaeng.applicationform.domain.event.OfficerRejectClubJoinApplicationEvent;
 import com.mohaeng.notification.domain.model.Notification;
 import com.mohaeng.notification.domain.model.NotificationMakeStrategy;
 import com.mohaeng.notification.domain.model.kind.OfficerRejectApplicationNotification;
 import com.mohaeng.notification.domain.model.value.Receiver;
+import com.mohaeng.participant.domain.event.OfficerRejectClubJoinApplicationEvent;
 import org.springframework.stereotype.Component;
 
 import java.util.List;
