+++ mohaeng/src/test/java/com/mohaeng/notification/application/eventhandler/NotificationEventHandlerAsyncTest.java
@@ -1,9 +1,9 @@
 package com.mohaeng.notification.application.eventhandler;
 
-import com.mohaeng.applicationform.domain.event.ClubJoinApplicationCreatedEvent;
 import com.mohaeng.common.annotation.ApplicationTest;
 import com.mohaeng.notification.domain.model.Notification;
 import com.mohaeng.notification.domain.repository.NotificationRepository;
+import com.mohaeng.participant.domain.event.ClubJoinApplicationCreatedEvent;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Nested;
 import org.junit.jupiter.api.Test;
